requirements.txt
streamlit
langchain>=0.0.350
langchain-groq>=0.0.6
langchain-community>=0.0.19
groq>=0.32.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pytesseract>=0.3.10
pillow>=10.0.0
pypdf>=3.12.0
tqdm>=4.65.0
numpy>=1.24.0
pathlib
easyocr
langchain-mistralai>=0.0.3
langchain-chroma